(CLEAN) Standardized type in BRV computation.

+++ master/libspider/spider/graphTransfo/CommonBRV.h
@@ -57,6 +57,6 @@ void updateBRV(transfoJob *job, long nVertices, int *brv, PiSDFVertex *const *ve
  * @param vertices  The vertices of the connected components.
  * @param nVertices The number of vertices inside the connected components.
  */
-void fillEdgeSet(PiSDFEdgeSet &edgeSet, PiSDFVertex *const *vertices, int nVertices);
+void fillEdgeSet(PiSDFEdgeSet &edgeSet, PiSDFVertex *const *vertices, long nVertices);
 
 #endif //SPIDER_COMMONBRV_H
